Tasks-Desktop: Refactor tasks sync code

+++ tasks-desktop/src/infrastructure/modules/tasks/repositories.js
@@ -1,6 +1,6 @@
 const { generateId } = require("../../../domain/shared/utils");
 const { Tables, Fields } = require("../../persistence/tables");
-const { Task, TaskSync } = require("../../../domain/modules/tasks");
+const { Task, TaskSync, TASK_SYNC_STATUS } = require("../../../domain/modules/tasks");
 
 const TasksRepository = (() => {
     const save = async (unitOfWork, task) => {
@@ -41,30 +41,62 @@ const TasksRepository = (() => {
     const get = async (unitOfWork, taskId) => {
         const queryManager = unitOfWork.getQueryManager();
         const result = await queryManager
-            .all(`SELECT *`
-                + `FROM ${Tables.Tasks} `
-                + `WHERE ${Fields.Tasks.Id} = ?`, [taskId]);
+            .all(`SELECT t.*, ts.synch_status `
+                + `FROM ${Tables.Tasks} t `
+                + `INNER JOIN ${Tables.TasksSync} ts ON ts.task_id = t.task_id `
+                + `WHERE t.${Fields.Tasks.Id} = ?`, [taskId]);
         return result 
             ? result.map((entity) => new Task(
                 entity.task_id
                 , entity.title
                 , entity.position
+                , entity.synch_status
                 , new Date(entity.created_at)
                 , new Date(entity.updated_at)
                 , entity.owner_id))[0] 
             : undefined;
     }
 
+    const getBySyncState = async (unitOfWork, syncStates = []) => {
+        const queryManager = unitOfWork.getQueryManager();
+        let preparedStatement = `SELECT t.*, ts.synch_status `
+            + `FROM ${Tables.Tasks} t `
+            + `INNER JOIN ${Tables.TasksSync} ts ON t.${Fields.Tasks.Id} = ts.${Fields.TasksSync.TaskId} `
+            + `WHERE ts.${Fields.TasksSync.Status} IN (` + syncStates.map(_ => '?').join(',') + `) `;
+
+        if (syncStates.filter(value => value === TASK_SYNC_STATUS.COMPLETE)) {
+            preparedStatement = `SELECT t.*, ts.synch_status `
+                + `FROM ${Tables.Tasks} t `
+                + `INNER JOIN ${Tables.TasksSync} ts on ts.task_id = t.task_id `
+                + `WHERE ${Fields.TasksSync.Status} IN (` + syncStates.map(_ => '?').join(',') + `) `;
+            
+        }
+        
+        const result = await queryManager.all(preparedStatement, syncStates);
+        return result.map(entry => new Task(
+            entry.task_id
+            , entry.title ? entry.title : ''
+            , entry.position ? entry.position : Infinity
+            , entry.synch_status
+            , entry.created_at ? new Date(entry.created_at) : null
+            , entry.updated_at ? new Date(entry.updated_at) : null
+            , entry.owner_id ? entry.owner_id : null
+        ));
+    }
+
     const listByAccountId = async (unitOfWork, accountId) => {
         const queryManager = unitOfWork.getQueryManager();
         const result = await queryManager
-            .all(`SELECT * FROM `
-                + `${Tables.Tasks} t `
-                + `WHERE ${Fields.Tasks.OwnerId} = ?`, [accountId])
+            .all(`SELECT t.*, ts.task_id as task_sync_task_id, ts.synch_status `
+                + `FROM ${Tables.Tasks} t `
+                + `INNER JOIN ${Tables.TasksSync} ts on ts.task_id = t.task_id `
+                + `WHERE ${Fields.Tasks.OwnerId} = ?`, [accountId]);
+
         return result.map(entry => new Task(
-                entry.task_id
+                entry.task_id != undefined ? entry.task_id : entry.task_sync_task_id
                 , entry.title
                 , entry.position
+                , entry.synch_status
                 , new Date(entry.created_at)
                 , new Date(entry.updated_at)
                 , entry.owner_id));
@@ -73,13 +105,16 @@ const TasksRepository = (() => {
     const listWithoutOwner = async (unitOfWork) => {
         const queryManager = unitOfWork.getQueryManager();
         const result = await queryManager
-            .all(`SELECT * `
-                + `FROM ${Tables.Tasks} `
+            .all(`SELECT t.*, ts.synch_status `
+                + `FROM ${Tables.Tasks} t `
+                + `RIGHT JOIN ${Tables.TasksSync} ts ON ts.task_id = t.task_id `
                 + `WHERE ${Fields.Tasks.OwnerId} is null`, []);
+
         return result.map(entry => new Task(
                 entry.task_id
                 , entry.title
                 , entry.position
+                , entry.synch_status
                 , new Date(entry.created_at)
                 , new Date(entry.updated_at)
                 , entry.owner_id));
@@ -88,13 +123,16 @@ const TasksRepository = (() => {
     const listById = async (unitOfWork, tasksIds = []) => {
         const queryManager = unitOfWork.getQueryManager();
         const result = await queryManager
-            .all(`SELECT * `
-                + `FROM ${Tables.Tasks} `
+            .all(`SELECT t.*, ts.synch_status `
+                + `FROM ${Tables.Tasks} t `
+                + `INNER JOIN ${Tables.TasksSync} ts ON t.task_id = ts.task_id `
                 + `WHERE ${Fields.Tasks.Id} in (` + tasksIds.map(task => '?').join(',') + ')', tasksIds);
+
         return result.map(entry => new Task(
                 entry.task_id
                 , entry.title
                 , entry.position
+                , entry.synch_status
                 , new Date(entry.created_at)
                 , new Date(entry.updated_at)
                 , entry.owner_id));
@@ -111,20 +149,27 @@ const TasksRepository = (() => {
 
     const erase = async (unitOfWork, id) => {
         const queryManager = unitOfWork.getQueryManager();
-        await queryManager.run(`DELETE FROM ${Tables.Tasks} `
-            + `WHERE ${Fields.Tasks.Id} = ?`, [id]);
+        const taskSyncs = await queryManager.all(`SELECT * FROM ${Tables.TasksSync} `
+            + `WHERE ${Fields.TasksSync.TaskId} = ?`, [id]);
+        if (taskSyncs.length == 0) {
+            await queryManager.run(`DELETE FROM ${Tables.Tasks} `
+                + `WHERE ${Fields.Tasks.Id} = ?`, [id]);
+        }
     }
 
     const eraseAllForAccount = async (unitOfWork, accountId) => {
         const queryManager = unitOfWork.getQueryManager();
-        await queryManager.run(`DELETE FROM ${Tables.Tasks} `
-            + `WHERE ${Fields.Tasks.OwnerId} = ?`, [accountId]);
+        await queryManager.run(`DELETE ` 
+            + `FROM ${Tables.Tasks} `
+            + `WHERE ${Fields.Tasks.OwnerId} = ?`
+            , [accountId]);
     }
 
     return {
         save
         , update
         , get
+        , getBySyncState
         , listByAccountId
         , listWithoutOwner
         , listById
@@ -193,8 +238,6 @@ const TasksSyncRepository = (() => {
 
     const update = async (unitOfWork, taskSyncData) => {
         const queryManager = unitOfWork.getQueryManager();
-
-        console.log(taskSyncData);
 
         let preparedStatement = `UPDATE ${Tables.TasksSync} `
             + `SET ${Fields.TasksSync.Status} = ?, ${Fields.TasksSync.UpdatedAt} = ? `
